+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Bug Fixes
 
+* Enbale globalReturn for commonjs environment
 * Use safe option for strict rule
 * Change wrap-iife setting to be compatible with prettier